fix: adding cast to float type in function thats set the number of elements

+++ src/main/java/FuzzyProject/FuzzyDT/Models/FDT.java
@@ -714,7 +714,7 @@ public class FDT {
         List<MicroGrupo> microGrupos = new ArrayList<>();
         for(int i=0; i<numGrupos; i++) {
             MicroGrupo mg = new MicroGrupo(dt);
-            mg.N = numElementosGrupo[i];
+            mg.N = (float) numElementosGrupo[i];
             microGrupos.add(mg);
         }
